+++ app/src/main/java/com/sembozdemir/revoluttest/main/MainModelMapper.kt
@@ -2,6 +2,7 @@ package com.sembozdemir.revoluttest.main
 
 import com.sembozdemir.revoluttest.core.network.model.CurrencyResponse
 import com.sembozdemir.revoluttest.core.network.model.Rates
+import java.math.BigDecimal
 import javax.inject.Inject
 import kotlin.reflect.full.memberProperties
 
@@ -19,7 +20,7 @@ class MainModelMapper @Inject constructor() {
     private fun mapCurrencyItems(rates: Rates): List<RateItem> {
         val currencyItems = arrayListOf<RateItem>()
         for (prop in Rates::class.memberProperties) {
-            (prop.get(rates) as? Double)?.let {
+            (prop.get(rates) as? BigDecimal)?.let {
                 currencyItems.add(
                     RateItem(
                         prop.name.toUpperCase(),
