+++ app/src/main/java/com/sembozdemir/revoluttest/main/RateItem.kt
@@ -1,6 +1,8 @@
 package com.sembozdemir.revoluttest.main
 
+import java.math.BigDecimal
+
 data class RateItem(
     val code: String,
-    val rate: Double
+    val rate: BigDecimal
 )
